+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Prevent ambassadors / GOALS-Admins to enter slots in stores that they are not a member of !4072 @AntonBallmaier
 - Disallow foodsavers to change their birthday #2183 !4075 @AntonBallmaier
 - New replying function in the chat component !4090 @alex.simm
+- Inverted action of the hygiene filter on the store team page !4127 @DL6ER
 
 ## Bugfixes
 - Resolve "region filter bug on profile page: Blank screen when navigating store list" #2295 !3940 @chriswalg